requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
